Move ramda to peer & dev dependencies
Ramda is required in order for the tests to run, so it needs to be included
as a devDep. Moving it to a peer dependency allows consumers to specify
which version of ramda to use without downloading extra versions. "0.x" as
the version constraint includes all the current 0.x versions, whereas "^0.19.0"
would only include versions in the provided minor range because ramda
hasn't hit 1.0 yet.

Fixes #17.